+++ factual/query/resolve.py
@@ -1,8 +1,8 @@
 from read import Read
 
 class Resolve(Read):
-    def __init__(self, api, values={}):
-        Read.__init__(self, api, 'places/resolve', values)
+    def __init__(self, api, table, values={}):
+        Read.__init__(self, api, 't/' + table + '/resolve', values)
 
     def values(self, values):
         return self._copy({'values': values})